requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
nltk>=3.8.1
pandas>=2.1.0
matplotlib>=3.8.0
streamlit>=1.28.0
yfinance>=0.2.36
scikit-learn>=1.4.0
seaborn>=0.13.0
finnhub-python>=2.4.18
python-dotenv>=0.19.0 